server/requirements.txt
mcp
python-jobspy
openai
python-dotenv
pandas
beautifulsoup4
